+++ src/components/FormContent/FormContent.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 
 import useStyles from './styles';
-import { updateTOC, cleanUpTOC } from '../Navigation/Navigation';
-import FormSectionHeader from '../FormSectionHeader';
-import IdentifiersSection from 'sections/IdentifiersSection';
+import { updateTOC, cleanUpTOC } from 'components/Navigation';
+import FormSectionHeader from 'components/FormSectionHeader';
+import IdentifiersSection from 'components/sections/IdentifiersSection';
 
 const sections = [
   { id: 'identifiers', title: 'identifiers', section: <IdentifiersSection /> },
