+++ src/components/Navigation/index.js
@@ -1,1 +1,2 @@
 export { default } from './Navigation';
+export { updateTOC, cleanUpTOC } from './Navigation';
